requirements.txt
requests-html==0.10.0
python-dotenv==0.21.1
boto3==1.26.72